com_scrape_types

Struct ComRef

pub struct ComRef<'a, I: Interface> { /* private fields */ }

A non-owning smart pointer to a COM object.

A ComRef<'a, I> represents a borrowed reference to a COM object implementing interface I. Like ComPtr, ComRef can be used to call interface methods on the referenced object. Unlike ComPtr, ComRef does not manage the objectโs reference count, i.e. it will not call the release method of the object it points to when going out of scope. See the crate-level documentation for more information.

A ComRef can be created safely from a ComPtr via ComPtr::as_com_ref, or from a ComWrapper via ComWrapper::as_com_ref. It can also be created unsafely via ComRef::from_raw.

Implementations

impl<'a, I: Interface> ComRef<'a, I>

pub fn as_ptr(&self) -> *mut I

Gets the wrapped interface pointer.

Does not perform any reference counting operations.

pub unsafe fn from_raw(ptr: *mut I) -> Option<ComRef<'a, I>>

Creates a ComRef from a raw interface pointer if the pointer is non-null.

Does not perform any reference counting operations.

from_raw will check if ptr is null (and return None if so), but this method is still unsafe, as the caller must still ensure that ptr is a valid interface pointer (see below) if it is non-null.

Safety

If ptr is non-null, it must be a valid pointer to a value of type I, and if *ptr is reinterpreted as *const I::Vtbl (see the safety documentation for Interface), it must in turn be a valid pointer to I::Vtbl.

pub unsafe fn from_raw_unchecked(ptr: *mut I) -> ComRef<'a, I>

Creates a ComRef from a raw interface pointer.

Does not perform any reference counting operations.

Safety

ptr must be a valid pointer to a value of type I, and if *ptr is reinterpreted as *const I::Vtbl (see the safety documentation for Interface), it must in turn be a valid pointer to I::Vtbl.

pub fn to_com_ptr(&self) -> ComPtr<I>

Upgrades the ComRef to a ComPtr.

Increments the reference count of the object that the ComRef points to.

pub fn upcast<J: Interface>(self) -> ComRef<'a, J>
where I: Inherits<J>,

Casts the ComRef from a derived interface to a base interface.

Does not perform any reference counting operations.

pub fn cast<J: Interface>(&self) -> Option<ComPtr<J>>

Attempts to cast from one interface to another, returning a ComPtr if successful.

If the cast is successful, increments the reference count of the object that the ComRef points to.