com_scrape_types

Trait Construct

pub unsafe trait Construct<C: Class, W: Wrapper<C>, const OFFSET: isize>: Interface {
    const OBJ: Self;
}

Generates the virtual table and base class object for a given class and interface.

The W parameter is the wrapper type which provides helper functionality for reference counting and class layout. The OFFSET parameter is the offset, within the header, at which the constructed base class object will be located.

This trait is used in the implementations of MakeHeader for tuples of interfaces. Interface types generated by com-scrape will implement this trait.

Safety

When I::OBJ is reinterpreted as *const I::Vtbl (see the safety documentation for Interface), it must be a valid pointer to an instance of I::Vtbl.

Required Associated Constants

const OBJ: Self

The generated base class object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.