com_scrape_types

Trait MakeHeader

pub unsafe trait MakeHeader<C, W>: InterfaceList
where
    C: Class,
    W: Wrapper<C>,
{
    const HEADER: Self::Header;
}

Generates the object header for a given class and list of interfaces.

This trait is implemented for tuples of interface types, and it is used by ComWrapper to construct the object header for a given Rust value.

Safety

If L::query(I::IID) returns Some(offset) for an Interface I (see InterfaceList), the I value located at offset offset within L::Header must (when reinterpreted as *const I::Vtbl) be a valid pointer to an instance of I::Vtbl.

Required Associated Constants

const HEADER: Self::Header

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.