com_scrape_types

Trait InterfaceList

pub unsafe trait InterfaceList {
    type Header;

    // Required method
    fn query(iid: &Guid) -> Option<isize>;
}

A list of COM interfaces implemented by a Rust type.

Provides a header type containing base class objects for each interface in the list, as well as a method for querying whether an interface is a member of the list and, if so, at what offset the corresponding base class object is located in the header.

This trait is implemented for tuples of interface types.

Safety

If L::query(I::IID) returns Some(offset) for an Interface I, then whenever ptr: *mut L::Header points to a valid instance of L::Header, (ptr as *mut u8).offset(offset) as *mut I must point to a valid instance of I.

Required Associated Types

type Header

Header type containing a base class object for each of the interfaces in the list.

Required Methods

fn query(iid: &Guid) -> Option<isize>

If there is an interface in the list whose GUID equals iid, or which transitively derives from an interface whose GUID equals iid, query returns the offset of the corresponding base class object within Self::Header.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.